добавила auth slice
