+++ client/src/store/store.ts
@@ -1,13 +1,16 @@
-import { configureStore } from '@reduxjs/toolkit';
-import userSlice from './features/userSlice';
+import { configureStore, MiddlewareArray } from '@reduxjs/toolkit';
+import authSlice from './features/authSlice';
 
 export const store = configureStore({
   reducer: {
-    user: userSlice,
+    auth: authSlice,
   },
+  // middleware: new MiddlewareArray().concat(additionalMiddleware, logger)
 });
 
 // Infer the `RootState` and `AppDispatch` types from the store itself
 export type RootState = ReturnType<typeof store.getState>
 // Inferred type: {posts: PostsState, comments: CommentsState, users: UsersState}
 export type AppDispatch = typeof store.dispatch
+
+export default store
